bpf: add hybrid NodePort mode (DSR for TCP, SNAT for UDP)
Add a mode where all TCP connections are processed via DSR and all UDP
connections via SNAT.

This allows to avoid any MTU changes in the cluster while still benefitting
from DSR for TCP when it's assumed the transport for the main workloads.
MTU does not need changes here since the service address is passed via IP
option in the TCP SYN only.

Example with UDP (SNAT) on node serving the NodePort request:

  14:22:50.891019 IP 192.168.178.30.43771 > 192.168.178.29.30001: UDP, length 44
  14:22:50.891036 IP 192.168.178.29.43771 > 10.12.227.188.69:  44 RRQ "hello" octet tsize 0 blksize 512 timeout 6
  14:22:50.891417 IP 10.12.227.188.69 > 192.168.178.29.43771:  24 OACK blksize 512 timeout 6
  14:22:50.891422 IP 192.168.178.29.30001 > 192.168.178.30.43771: UDP, length 24
  [...]

Example with TCP (DSR) on node serving the NodePort request:

  14:23:09.142617 IP 192.168.178.30.43094 > 192.168.178.29.30000: Flags [S], seq 1481227239, win 64240, options [mss 1460,sackOK,TS val 115607409 ecr 0,nop,wscale 7], length 0
  14:23:09.142631 IP 192.168.178.30.43094 > 10.12.94.202.80: Flags [S], seq 1481227239, win 64240, options [mss 1460,sackOK,TS val 115607409 ecr 0,nop,wscale 7], length 0
  14:23:09.145232 IP 192.168.178.30.43094 > 192.168.178.29.30000: Flags [.], ack 1603130306, win 502, options [nop,nop,TS val 115607411 ecr 2597942183], length 0
  14:23:09.145236 IP 192.168.178.30.43094 > 10.12.94.202.80: Flags [.], ack 1603130306, win 502, options [nop,nop,TS val 115607411 ecr 2597942183], length 0
  14:23:09.145237 IP 192.168.178.30.43094 > 192.168.178.29.30000: Flags [P.], seq 0:84, ack 1, win 502, options [nop,nop,TS val 115607411 ecr 2597942183], length 84
  14:23:09.145239 IP 192.168.178.30.43094 > 10.12.94.202.80: Flags [P.], seq 0:84, ack 1, win 502, options [nop,nop,TS val 115607411 ecr 2597942183], length 84: HTTP: GET / HTTP/1.1
  [...]

As can be seen in the UDP case above, we need to perform the reverse
DNAT/SNAT from the replies while in the TCP case we only ever see the
requests.

Tested on v4.19.57 kernel.

Signed-off-by: Daniel Borkmann <daniel@iogearbox.net>